refactor: create src directory in client Mentions: #13
