+++ client/android/app/src/main/AndroidManifest.xml
@@ -28,6 +28,7 @@
         <category android:name="android.intent.category.BROWSABLE"/>
         <data android:scheme="com.hiroto0222.workouttracker"/>
         <data android:scheme="exp+workout-tracker"/>
+        <data android:scheme="workout-tracker"/>
       </intent-filter>
     </activity>
     <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" android:exported="false"/>
