+++ client/package.json
@@ -1,6 +1,7 @@
 {
   "name": "workout-tracker",
   "version": "1.0.0",
+  "main": "src/App.tsx",
   "scripts": {
     "start": "expo start --dev-client",
     "android": "expo run:android",
@@ -32,6 +33,7 @@
     "@types/react": "~18.0.14",
     "@types/react-native": "^0.72.2",
     "@types/react-native-dotenv": "^0.2.0",
+    "babel-plugin-module-resolver": "^5.0.0",
     "typescript": "^4.9.4"
   },
   "private": true
