+++ client/src/App.tsx
@@ -1,9 +1,10 @@
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { registerRootComponent } from "expo";
 import { StyleSheet } from "react-native";
 import { Provider } from "react-redux";
-import HomeScreen from "./screens/HomeScreen";
-import LoginScreen from "./screens/LoginScreen";
+import HomeScreen from "./components/screens/HomeScreen";
+import LoginScreen from "./components/screens/LoginScreen";
 import { store } from "./store";
 
 const Stack = createNativeStackNavigator();
@@ -24,6 +25,8 @@ export default function App() {
     </NavigationContainer>
   );
 }
+
+registerRootComponent(App);
 
 const styles = StyleSheet.create({
   container: {
