+++ client/src/assets/splash.png
