+++ client/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
-    "strict": true
+    "strict": true,
+    "baseUrl": "./src"
   },
-  "typeRoots": ["./types"]
+  "typeRoots": ["./src/types"]
 }
